Add subscriptions package
