+++ packages/effection/package.json
@@ -1,6 +1,6 @@
 {
   "name": "effection",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "license": "MIT",
