+++ packages/events/package.json
@@ -20,8 +20,8 @@
     "mocha": "mocha -r ts-node/register"
   },
   "dependencies": {
-    "effection": "^0.7.0",
-    "@effection/subscription": "^0.11.0"
+    "effection": "^2.0.0",
+    "@effection/subscription": "^2.0.0"
   },
   "devDependencies": {
     "@frontside/tsconfig": "0.0.1",
