+++ packages/subscription/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@effection/subscription",
-  "version": "0.11.0",
+  "version": "2.0.0",
   "description": "Effection Subscriptions",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -19,10 +19,11 @@
     "mocha": "mocha -r ts-node/register"
   },
   "dependencies": {
-    "effection": "^0.7.0"
+    "effection": "^2.0.0"
   },
   "devDependencies": {
     "@frontside/tsconfig": "0.0.1",
+    "@types/mocha": "^8.0.3",
     "expect": "^25.4.0",
     "mocha": "^7.1.1",
     "ts-node": "^8.9.0",
