+++ packages/subscription/src/operation-iterable.ts
@@ -1,0 +1,7 @@
+import { Operation, Task } from 'effection';
+import { OperationIterator } from './operation-iterator';
+import { SymbolOperationIterable } from './symbol-operation-iterable';
+
+export interface OperationIterable<T,TReturn> {
+  [SymbolOperationIterable](task: Task<unknown>): OperationIterator<T,TReturn>;
+}
