+++ packages/subscription/src/operation-iterator.ts
@@ -1,0 +1,5 @@
+import { Operation } from 'effection';
+
+export interface OperationIterator<T,TReturn> {
+  next(): Operation<IteratorResult<T,TReturn>>;
+}
